test(smoketest): move into package smoketest
Switch from package smoketest_test to package smoketest so the test
file follows the convention of living in the same package as the code
under test. The cmd/smoketest directory contains no production source
file; the test binary continues to compile and run as before.
