+++ packages/orchestrator/cmd/smoketest/smoke_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package smoketest_test
+package smoketest
 
 import (
 	"context"
